Move bloom filter classes from obsolete bloom_filter.hpp to new tablefilter_internal_functions
