+++ src/include/duckdb/execution/join_hashtable.hpp
@@ -19,8 +19,8 @@
 #include "duckdb/common/unique_ptr.hpp"
 #include "duckdb/execution/aggregate_hashtable.hpp"
 #include "duckdb/execution/ht_entry.hpp"
-#include "duckdb/planner/filter/bloom_filter.hpp"
 #include "duckdb/planner/filter/prefix_range_filter.hpp"
+#include "duckdb/planner/filter/tablefilter_internal_functions.hpp"
 
 namespace duckdb {
