+++ src/include/duckdb/planner/filter/tablefilter_internal_functions.hpp
@@ -19,9 +19,32 @@
 namespace duckdb {
 
 class BaseStatistics;
-class BloomFilter;
 class PerfectHashJoinExecutor;
 class PrefixRangeFilter;
+
+class BloomFilter {
+public:
+	BloomFilter() = default;
+	void Initialize(ClientContext &context_p, idx_t number_of_rows);
+
+	void InsertHashes(const Vector &hashes_v, idx_t count) const;
+	idx_t LookupHashes(const Vector &hashes_v, SelectionVector &result_sel, idx_t count) const;
+
+	void InsertOne(hash_t hash) const;
+	bool LookupOne(hash_t hash) const;
+
+	bool IsInitialized() const {
+		return initialized;
+	}
+
+private:
+	idx_t num_sectors;
+	uint64_t bitmask; // num_sectors - 1 -> used to get the sector offset
+
+	bool initialized = false;
+	AllocatedData buf_;
+	uint64_t *bf;
+};
 
 struct DynamicFilterData {
 	DynamicFilterData() : comparison_type(ExpressionType::INVALID) {
